# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Relationship": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_person_id": {
            "format": "uuid",
            "title": "From Person Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "qualifier": {
            "title": "Qualifier",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_person_id": {
            "format": "uuid",
            "title": "To Person Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RelationshipType"
          }
        },
        "required": [
          "id",
          "space_id",
          "type",
          "from_person_id",
          "to_person_id",
          "qualifier",
          "status",
          "created_at"
        ],
        "title": "Relationship",
        "type": "object"
      },
      "RelationshipCreate": {
        "additionalProperties": false,
        "description": "For `parent_child`, `from_person_id` is the parent and `to_person_id` the child.",
        "properties": {
          "from_person_id": {
            "format": "uuid",
            "title": "From Person Id",
            "type": "string"
          },
          "qualifier": {
            "title": "Qualifier",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UnionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_person_id": {
            "format": "uuid",
            "title": "To Person Id",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/RelationshipType"
          }
        },
        "required": [
          "type",
          "from_person_id",
          "to_person_id",
          "qualifier"
        ],
        "title": "RelationshipCreate",
        "type": "object"
      },
      "RelationshipType": {
        "enum": [
          "parent_child",
          "union"
        ],
        "title": "RelationshipType",
        "type": "string"
      },
...
      "UnionStatus": {
        "enum": [
          "active",
          "divorced",
          "separated",
          "widowed",
          "annulled"
        ],
        "title": "UnionStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Evidence-based family trees and family memory. AGPL-3.0-only.",
    "title": "family-history API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/relationships/{relationship_id}": {
      "delete": {
        "operationId": "delete_relationship_v1_relationships__relationship_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "relationship_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Relationship Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Delete Relationship",
        "tags": [
          "relationships"
        ]
      }
    },
...
    "/v1/spaces/{space_id}/relationships": {
      "post": {
        "description": "For `parent_child`, `from_person_id` is the parent and `to_person_id` the child.",
        "operationId": "create_relationship_v1_spaces__space_id__relationships_post",
        "parameters": [
          {
            "in": "path",
            "name": "space_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Space Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Relationship"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Create Relationship",
        "tags": [
          "relationships"
        ]
      }
    },
...
  }
}